+++ packages/ui/textarea/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/textarea
 
+## 3.3.9
+
+### Patch Changes
+
+- Remove dependencies
+
 ## 3.3.8
 
 ### Patch Changes